[mp4tree] Break out NAL and SEI parsing
  * A refactoring step since main file has grown a bit
