+++ common.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <stdlib.h>
+#include <stdint.h>
+
+const char *
+indent(int depth, int header);
